# api/openapi.yaml
openapi: 3.0.3
info:
  title: FQP From Scratch API
  version: 1.0.0
  description: 官方足彩/竞彩足球量化研究与长期运行系统 API 草案。系统只提供数据分析、模拟与复盘接口，不提供购彩、代购、出票、收款接口。
servers:
  - url: http://localhost:8000
paths:
  /api/v1/official/matches:
    get:
      summary: 查询官方赛程
      parameters:
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: 官方赛程列表
  /api/v1/official/matches/{match_id}:
    get:
      summary: 查询比赛详情
      parameters:
        - name: match_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 比赛详情 }
  /api/v1/official/matches/{match_id}/odds-snapshots:
    get:
      summary: 查询赔率快照
      parameters:
        - name: match_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 赔率快照列表 }
...
  /api/v1/pool/analyze:
    get:
      summary: 运行当前实现的传统足彩分析
      parameters:
        - name: budget
          in: query
          required: false
          schema: { type: integer, default: 256 }
        - name: strategy
          in: query
          required: false
          schema:
            type: string
            enum: [conservative, balanced, aggressive]
            default: balanced
      responses:
        '200': { description: 传统足彩分析结果 }
  /api/v1/pool/issues/{id}/generate-combinations:
    post:
      summary: 生成传统足彩组合覆盖方案
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: 组合方案 }
...